HID connect and query out of application and into library.
