+++ VS/LewanSoul-xArm/Consoles/USB/MainPage.xaml.cs
@@ -1,22 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Devices.Enumeration;
-using Windows.Devices.HumanInterfaceDevice;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.Storage;
-using Windows.Storage.Streams;
-using Windows.UI.Core;
-using Windows.UI.Xaml;
+﻿using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
+using xArmDotNet;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
 
@@ -29,66 +14,9 @@ namespace xArmUSBConsole
     {
         public MainPage()
         {
-            this.InitializeComponent();
+            InitializeComponent();
 
-            EnumerateHidDevicesAsync();
-        }
-
-        private async void EnumerateHidDevicesAsync()
-        {
-            ushort vendorId = 0x0483;
-            ushort productId = 0x5750;
-            ushort usagePage = 0x008C;
-            ushort usageId = 0x0001;
-
-            string selector = HidDevice.GetDeviceSelector(usagePage, usageId, vendorId, productId);
-            var devices = await DeviceInformation.FindAllAsync(selector);
-
-            if (devices.Any())
-            {
-                // At this point the device is available to communicate with
-                // So we can send/receive HID reports from it or 
-                // query it for control descriptions.
-                info.Text = "HID devices found: " + devices.Count;
-
-                foreach (var item in devices.ElementAt(0).Properties.ToList())
-                {
-                    info.Text += "\n   " + item.Key.ToString() + ": " + Convert.ToString(item.Value);
-                }
-
-                // Open the target HID device.
-                HidDevice device = await HidDevice.FromIdAsync(devices.ElementAt(0).Id, FileAccessMode.ReadWrite);
-                
-
-                if (device != null)
-                {
-                    info.Text += "\nInstalling report handler.";
-
-                    // Input reports contain data from the device.
-                    device.InputReportReceived += async (sender, args) =>
-                    {
-                        HidInputReport inputReport = args.Report;
-                        IBuffer buffer = inputReport.Data;
-
-                        // Create a DispatchedHandler as we are interracting with the UI directly and the
-                        // thread that this function is running on might not be the UI thread; 
-                        // if a non-UI thread modifies the UI, an exception is thrown.
-
-                        await Dispatcher.RunAsync(
-                            CoreDispatcherPriority.Normal,
-                            new DispatchedHandler(() =>
-                            {
-                                info.Text += "\nHID Input Report: " + inputReport.ToString() +
-                                "\nTotal number of bytes received: " + buffer.Length.ToString();
-                            }));
-                    };
-                }
-            }
-            else
-            {
-                // There were no HID devices that met the selector criteria.
-                info.Text = "HID device not found";
-            }
+            Robot robot = new Robot();
         }
 
         // Causes info TextBox to scroll to bottom when updated.
