+++ VS/LewanSoul-xArm/Library/xArmDotNet/xArmDotNet.csproj
@@ -106,7 +106,7 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
+    <Compile Include="Robot.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\xArmDotNet.rd.xml" />
   </ItemGroup>
